Convert tests to a custom test harness
This is the only way we can create the `EventLoop` multiple times
throughout the test suite, as there can only be one instance per
program.

As a nice side-effect, this also removes the need for platform-specific,
"any-thread" setup, so the tests can now run on all supported platforms.
