+++ crates/wolf_engine_window/Cargo.toml
@@ -27,3 +27,14 @@ anyhow = "1.0.86"
 
 [dev-dependencies]
 ntest = "0.9.3"
+libtest-mimic = "0.7.3"
+
+[[test]]
+name = "run_and_quit"
+path = "tests/run_and_quit.rs"
+harness = false
+
+[[test]]
+name = "dropped_window"
+path = "tests/dropped_window.rs"
+harness = false
