+++ crates/wolf_engine_window/src/event.rs
@@ -55,7 +55,7 @@ impl EventLoopBuilder {
     /// Initialize the window system.
     pub fn build(self) -> Result<EventLoop, WindowError> {
         match WinitEventLoop::with_user_event().build() {
-            Ok(event_loop) => Ok(self.build_with_event_loop(event_loop)),
+            Ok(event_loop) => Ok(EventLoop::new(event_loop)),
             Err(error) => match error {
                 EventLoopError::Os(error) => {
                     Err(WindowError::OsError(OsError::from(anyhow!(error))))
@@ -67,11 +67,6 @@ impl EventLoopBuilder {
                 error => panic!("Unhandled Error: {error}"),
             },
         }
-    }
-
-    #[allow(deprecated)]
-    pub(crate) fn build_with_event_loop(self, event_loop: WinitEventLoop) -> EventLoop {
-        EventLoop::new(event_loop)
     }
 }
